fix: resolve all clippy warnings in asterisk-codecs crate
- Remove unnecessary parentheses in tone_gen.rs
- Add #[allow(clippy::same_item_push)] for intentional zero-push loop
- Convert .max(X).min(Y) to .clamp(X, Y) in g722.rs, g726.rs, adpcm.rs
- Add #[allow(clippy::needless_range_loop)] for DSP indexing loops
  in g722.rs, noise_suppress.rs, plc.rs
- Add #[allow(clippy::should_implement_trait)] for opus.rs from_str
- Replace manual if/else clamping with .clamp() in adpcm.rs

+++ crates/asterisk-codecs/src/tone_gen.rs
@@ -204,10 +204,11 @@ impl ToneGenerator {
 
             if self.in_silence {
                 // Generate silence
-                let silence_samples = (segment.silence_ms * self.sample_rate / 1000) as u32;
+                let silence_samples = segment.silence_ms * self.sample_rate / 1000;
                 let remaining = silence_samples.saturating_sub(self.sample_position);
                 let to_generate = remaining.min((num_samples - output.len()) as u32);
 
+                #[allow(clippy::same_item_push)]
                 for _ in 0..to_generate {
                     output.push(0);
                 }
@@ -221,7 +222,7 @@ impl ToneGenerator {
                 }
             } else {
                 // Generate tone
-                let tone_samples = (segment.duration_ms * self.sample_rate / 1000) as u32;
+                let tone_samples = segment.duration_ms * self.sample_rate / 1000;
                 let remaining = tone_samples.saturating_sub(self.sample_position);
                 let to_generate = remaining.min((num_samples - output.len()) as u32);
